VERY EXPERIMENTAL CHANGES to libseq data types.  The original authors didn't define their integer types, resulting in 64-bit ints where the program should have 32-bit ints, etc.  This will remedy that problem, but apparently causes some major runtime bugs.
git-svn-id: http://libseq-branch.googlecode.com/svn/trunk@4 419bbb26-8e0e-33f9-3231-470fb831da05

+++ own_types.h
@@ -32,12 +32,12 @@ using namespace std;
 /////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////
 
-typedef unsigned short      US_int;  // 16 bit without sign    //
-typedef unsigned long       UL_int;  // 32 bit without sign    //
-typedef unsigned long long ULL_int;  // 64 bit without sign    //
-typedef          short       S_int;  // 16 bit with sign       //
-typedef          long        L_int;  // 32 bit with sign       //
-typedef          long long  LL_int;  // 64 bit with sign       //
+typedef uint16_t      US_int;  // 16 bit without sign    //
+typedef uint32_t       UL_int;  // 32 bit without sign    //
+typedef uint64_t ULL_int;  // 64 bit without sign    //
+typedef          int16_t       S_int;  // 16 bit with sign       //
+typedef          int32_t        L_int;  // 32 bit with sign       //
+typedef          int64_t  LL_int;  // 64 bit with sign       //
 
 typedef UL_int             R_Elem;   // ring elements          //
 typedef UL_int             FixPoint; // fixed point reals      //
